It turns out that, in gcc,-Wall does actually mean 'all'. Using the
flags taken from a discussion on stack overflow seems like over kill,
but I will keep a few of them that appear useful in stopping lazy
coding.

The individual commits will show what was altered to achive
warning-less compliation.

A bug was caught when writing the EPS key for ground-state half life.

+++ src/convertFloatToExponent.cpp
@@ -1,6 +1,6 @@
 #include "include/functions.h"
 
-void convertFloatToExponent(const float &in,
+void convertFloatToExponent(const double &in,
 			    std::vector<std::string> &out
 			    )
 {
